+++ app/src/main/java/com/kirchhoff/movies/di/NetworkModule.kt
@@ -5,8 +5,8 @@ import com.kirchhoff.movies.BuildConfig
 import com.kirchhoff.movies.network.interceptors.ApiKeyInterceptor
 import com.kirchhoff.movies.network.services.DiscoverService
 import com.kirchhoff.movies.network.services.MovieService
-import com.kirchhoff.movies.network.services.TvService
 import com.kirchhoff.movies.screen.person.network.PersonService
+import com.kirchhoff.movies.screen.tvshow.network.TvShowService
 import okhttp3.OkHttpClient
 import org.koin.dsl.module
 import retrofit2.Retrofit
@@ -34,5 +34,5 @@ val networkModule = module {
 
     single { get<Retrofit>().create(MovieService::class.java) }
 
-    single { get<Retrofit>().create(TvService::class.java) }
+    single { get<Retrofit>().create(TvShowService::class.java) }
 }
