--- app/src/main/java/com/kirchhoff/movies/repository/tv/ITvRepository.kt
@@ -1,10 +1,0 @@
-package com.kirchhoff.movies.repository.tv
-
-import com.kirchhoff.movies.core.data.UIEntertainmentCredits
-import com.kirchhoff.movies.core.repository.Result
-import com.kirchhoff.movies.data.ui.details.tv.UITvDetails
-
-interface ITvRepository {
-    suspend fun fetchDetails(tvId: Int): Result<UITvDetails>
-    suspend fun fetchTvCredits(tvId: Int): Result<UIEntertainmentCredits>
-}
