+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/MainPagerAdapter.kt
@@ -4,8 +4,8 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
 import com.kirchhoff.movies.screen.person.ui.screen.list.PersonsFragment
+import com.kirchhoff.movies.screen.tvshow.ui.screen.list.TvShowListFragment
 import com.kirchhoff.movies.ui.screens.main.movies.MoviesFragment
-import com.kirchhoff.movies.ui.screens.main.tvs.TvsFragment
 
 class MainPagerAdapter(fm: FragmentManager) :
     FragmentPagerAdapter(fm, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT) {
@@ -13,7 +13,7 @@ class MainPagerAdapter(fm: FragmentManager) :
     override fun getItem(position: Int): Fragment {
         return when (position) {
             0 -> MoviesFragment()
-            1 -> TvsFragment()
+            1 -> TvShowListFragment()
             else -> PersonsFragment()
         }
     }
