--- app/src/main/java/com/kirchhoff/movies/ui/screens/main/tvs/TvsVM.kt
@@ -1,10 +1,0 @@
-package com.kirchhoff.movies.ui.screens.main.tvs
-
-import com.kirchhoff.movies.core.data.UITv
-import com.kirchhoff.movies.core.ui.paginated.PaginatedScreenVM
-import com.kirchhoff.movies.core.ui.paginated.UIPaginated
-import com.kirchhoff.movies.repository.discover.IDiscoverRepository
-
-class TvsVM(private val discoverRepository: IDiscoverRepository) : PaginatedScreenVM<UIPaginated<UITv>>() {
-    override suspend fun loadData(page: Int) = discoverRepository.fetchTvs(page)
-}
