+++ screen/tvshow/src/main/java/com/kirchhoff/movies/screen/tvshow/data/UITvShowDetails.kt
@@ -1,0 +1,14 @@
+package com.kirchhoff.movies.screen.tvshow.data
+
+import com.kirchhoff.movies.core.data.UIGenre
+
+data class UITvShowDetails(
+    val numberOfSeasons: Int,
+    val numberOfEpisodes: Int,
+    val overview: String,
+    val status: String,
+    val firstAirDate: String,
+    val voteCount: Int?,
+    val voteAverage: Float?,
+    val genres: List<UIGenre>
+)
