+++ screen/tvshow/src/main/java/com/kirchhoff/movies/screen/tvshow/repository/ITvShowRepository.kt
@@ -1,0 +1,13 @@
+package com.kirchhoff.movies.screen.tvshow.repository
+
+import com.kirchhoff.movies.core.data.UIEntertainmentCredits
+import com.kirchhoff.movies.core.data.UITv
+import com.kirchhoff.movies.core.repository.Result
+import com.kirchhoff.movies.core.ui.paginated.UIPaginated
+import com.kirchhoff.movies.screen.tvshow.data.UITvShowDetails
+
+interface ITvShowRepository {
+    suspend fun fetchDiscoverList(page: Int): Result<UIPaginated<UITv>>
+    suspend fun fetchDetails(tvId: Int): Result<UITvShowDetails>
+    suspend fun fetchCredits(tvId: Int): Result<UIEntertainmentCredits>
+}
