+++ screen/tvshow/src/main/java/com/kirchhoff/movies/screen/tvshow/ui/screen/list/TvShowListFragment.kt
@@ -1,0 +1,34 @@
+package com.kirchhoff.movies.screen.tvshow.ui.screen.list
+
+import com.kirchhoff.movies.core.data.UITv
+import com.kirchhoff.movies.core.ui.paginated.PaginatedScreenFragment
+import com.kirchhoff.movies.core.ui.paginated.UIPaginated
+import com.kirchhoff.movies.core.ui.recyclerview.adapter.BaseRecyclerViewAdapter
+import com.kirchhoff.movies.screen.tvshow.R
+import com.kirchhoff.movies.screen.tvshow.ui.screen.list.adapter.TvShowListAdapter
+import org.koin.android.viewmodel.ext.android.viewModel
+
+class TvShowListFragment : PaginatedScreenFragment<UITv, UIPaginated<UITv>>(),
+    BaseRecyclerViewAdapter.OnItemClickListener<UITv> {
+
+    override val vm by viewModel<TvShowListVM>()
+
+    override val listAdapter = TvShowListAdapter(this)
+
+    override val configuration: Configuration = Configuration(
+        threshold = THRESHOLD,
+        spanCount = SPAN_COUNT,
+        emptyResultText = R.string.empty_tw_shows,
+        isToolbarVisible = false,
+        toolbarTitle = ""
+    )
+
+    companion object {
+        private const val SPAN_COUNT = 1
+        private const val THRESHOLD = 3
+    }
+
+    override fun onItemClick(item: UITv) {
+        router.openTvDetailsScreen(item)
+    }
+}
