+++ screen/tvshow/src/main/res/values/dimens.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="tv_details_poster_width">120dp</dimen>
+    <dimen name="tv_details_poster_height">175dp</dimen>
+</resources>
